chore(folio): extract comment state into useFolioComments hook
Move DocxEditor's comment subsystem (thread storage, author-visibility
filtering, the add-comment flow, and the DOM highlight sync) into a
dedicated useFolioComments hook. Save-coupled mutators stay in the
component since they depend on its document-build pipeline.

Behavior-preserving: the hook's return is destructured into the same
local names, so existing call sites are unchanged. First step in
decomposing the ~4k-line editor component; net -135 lines there.
