+++ packages/folio/src/components/DocxEditor.tsx
@@ -185,7 +185,6 @@ import {
   createComment,
   findSelectionYPosition,
   getCommentAuthorKey,
-  getCommentParentId,
   getFallbackCommentYPosition,
   pruneOrphanedComments,
   removePendingCommentMarkRange,
@@ -245,6 +244,7 @@ import { getBuiltinTableStyle } from "./ui/table-styles";
 import type { TableStylePreset } from "./ui/table-styles";
 import type { TableAction } from "./ui/table-types";
 import { Tooltip } from "./ui/Tooltip";
+import { useFolioComments } from "./useFolioComments";
 
 const CommentsSidebar = lazy(() =>
   import("./CommentsSidebar").then((m) => ({
@@ -497,28 +497,11 @@ export function DocxEditor({
   showOutlineRef.current = showOutline;
   const [outlineHeadings, setHeadingInfos] = useState<HeadingInfo[]>([]);
 
-  // Comments sidebar state
-  const [showCommentsSidebar, setShowCommentsSidebar] = useState(false);
-  const [visibleCommentAuthors, setVisibleCommentAuthors] =
-    useState<Set<string> | null>(null);
-  const [activeCommentId, setActiveCommentId] = useState<number | null>(null);
-  const [comments, setComments] = useState<Comment[]>([]);
-  const commentsDirtyRef = useRef(false);
-  const commentsRef = useRef<Comment[]>([]);
-  commentsRef.current = comments;
   const [, setTrackedChanges] = useState<TrackedChangeEntry[]>([]);
   const [anchorPositions, setAnchorPositions] = useState<Map<string, number>>(
     EMPTY_ANCHOR_POSITIONS,
   );
 
-  const [isAddingComment, setIsAddingComment] = useState(false);
-  const [commentSelectionRange, setCommentSelectionRange] = useState<{
-    from: number;
-    to: number;
-  } | null>(null);
-  const [addCommentYPosition, setAddCommentYPosition] = useState<number | null>(
-    null,
-  );
   const {
     editingMode,
     readOnly,
@@ -531,14 +514,6 @@ export function DocxEditor({
     onModeChange,
     readOnlyProp,
   });
-
-  // Floating "add comment" button position (relative to scroll container, null = hidden)
-  const [floatingCommentBtn, setFloatingCommentBtn] = useState<{
-    top: number;
-    left: number;
-    from: number;
-    to: number;
-  } | null>(null);
 
   // Debounce timer for extractTrackedChanges (avoid full doc walk on every keystroke)
   const extractTrackedChangesTimerRef = useRef<ReturnType<
@@ -630,89 +605,6 @@ export function DocxEditor({
     groupingInterval: 500,
     enableKeyboardShortcuts: true,
   });
-
-  // Extract comments from document model on initial load
-  const commentsLoadedRef = useRef(false);
-  useEffect(() => {
-    if (commentsLoadedRef.current) {
-      return;
-    }
-    const doc = history.state;
-    if (!doc) {
-      return;
-    }
-    const bodyComments = doc.package.document.comments;
-    if (bodyComments && bodyComments.length > 0) {
-      setComments(bodyComments);
-      setVisibleCommentAuthors(null);
-      setActiveCommentId(null);
-      if (autoOpenReviewSidebar) {
-        setShowCommentsSidebar(true);
-      }
-      commentsLoadedRef.current = true;
-    }
-  }, [autoOpenReviewSidebar, history.state]);
-
-  const commentAuthors = useMemo(() => {
-    const seen = new Set<string>();
-    const authors: string[] = [];
-    for (const comment of comments) {
-      const commentAuthor = getCommentAuthorKey(comment.author);
-      if (!seen.has(commentAuthor)) {
-        seen.add(commentAuthor);
-        authors.push(commentAuthor);
-      }
-    }
-    return authors;
-  }, [comments]);
-
-  const visibleCommentAuthorSet = useMemo(
-    () => visibleCommentAuthors ?? new Set(commentAuthors),
-    [visibleCommentAuthors, commentAuthors],
-  );
-
-  const visibleCommentIds = useMemo(() => {
-    const ids = new Set<number>([PENDING_COMMENT_ID]);
-    for (const comment of comments) {
-      if (visibleCommentAuthorSet.has(getCommentAuthorKey(comment.author))) {
-        ids.add(comment.id);
-      }
-    }
-    return ids;
-  }, [comments, visibleCommentAuthorSet]);
-
-  const visibleComments = useMemo(() => {
-    const visibleRootIds = new Set<number>();
-    for (const comment of comments) {
-      const parentId = getCommentParentId(comment);
-      if (
-        parentId === null ||
-        parentId === undefined ||
-        !visibleCommentIds.has(comment.id)
-      ) {
-        continue;
-      }
-      visibleRootIds.add(parentId);
-    }
-    return comments.filter((comment) => {
-      const parentId = getCommentParentId(comment);
-      if (parentId !== null && parentId !== undefined) {
-        return visibleCommentIds.has(comment.id);
-      }
-      return (
-        visibleCommentIds.has(comment.id) || visibleRootIds.has(comment.id)
-      );
-    });
-  }, [comments, visibleCommentIds]);
-
-  const activeCommentVisible =
-    activeCommentId !== null && visibleCommentIds.has(activeCommentId);
-
-  useEffect(() => {
-    if (!activeCommentVisible) {
-      setActiveCommentId(null);
-    }
-  }, [activeCommentVisible]);
 
   // Extension manager — built once, provides schema + plugins + commands
   const extensionManager = useMemo(() => {
@@ -900,62 +792,35 @@ export function DocxEditor({
     color: { rgb: "000000" },
   });
 
-  const syncCommentHighlightStyles = useCallback(() => {
-    const root = editorContentRef.current;
-    if (!root) {
-      return;
-    }
-
-    const nodes = root.querySelectorAll<HTMLElement>(
-      ".layout-run-text[data-comment-id]",
-    );
-    for (const node of nodes) {
-      const commentId = Number.parseInt(node.dataset["commentId"] ?? "", 10);
-      const isPending = commentId === PENDING_COMMENT_ID;
-      const isVisible = isPending || visibleCommentIds.has(commentId);
-      if (!isVisible) {
-        node.style.backgroundColor = "transparent";
-        node.style.borderBottom = "2px solid transparent";
-        node.style.boxShadow = "none";
-        delete node.dataset["activeComment"];
-        continue;
-      }
-
-      if (activeCommentId === commentId) {
-        node.style.backgroundColor =
-          "var(--doc-comment-active-bg, rgba(255, 212, 0, 0.22))";
-        node.style.borderBottom =
-          "1px solid var(--doc-comment-active-border, rgba(180, 130, 0, 0.62))";
-        node.style.boxShadow = "none";
-        node.dataset["activeComment"] = "true";
-        continue;
-      }
-
-      node.style.backgroundColor =
-        "var(--doc-comment-bg, rgba(255, 212, 0, 0.08))";
-      node.style.borderBottom =
-        "1px solid var(--doc-comment-border, rgba(180, 130, 0, 0.24))";
-      node.style.boxShadow = "none";
-      delete node.dataset["activeComment"];
-    }
-  }, [visibleCommentIds, activeCommentId]);
-
-  useLayoutEffect(() => {
-    syncCommentHighlightStyles();
-  }, [syncCommentHighlightStyles, anchorPositions]);
-
-  useEffect(() => {
-    syncCommentHighlightStyles();
-    const firstFrame = requestAnimationFrame(() => {
-      syncCommentHighlightStyles();
-      requestAnimationFrame(syncCommentHighlightStyles);
-    });
-    const timeout = setTimeout(syncCommentHighlightStyles, 120);
-    return () => {
-      cancelAnimationFrame(firstFrame);
-      clearTimeout(timeout);
-    };
-  }, [comments.length, syncCommentHighlightStyles]);
+  const {
+    comments,
+    setComments,
+    commentsRef,
+    commentsDirtyRef,
+    commentsLoadedRef,
+    showCommentsSidebar,
+    setShowCommentsSidebar,
+    setVisibleCommentAuthors,
+    activeCommentId,
+    setActiveCommentId,
+    isAddingComment,
+    setIsAddingComment,
+    commentSelectionRange,
+    setCommentSelectionRange,
+    addCommentYPosition,
+    setAddCommentYPosition,
+    floatingCommentBtn,
+    setFloatingCommentBtn,
+    commentAuthors,
+    visibleCommentAuthorSet,
+    visibleComments,
+    syncCommentHighlightStyles,
+  } = useFolioComments({
+    doc: history.state,
+    autoOpenReviewSidebar,
+    anchorPositions,
+    editorContentRef,
+  });
   // Cache style resolver to avoid recreating on every selection change
   const styleResolverCacheRef = useRef<{
     styles: unknown;
